Fix gpushd-server

+++ safe-call.h
@@ -32,6 +32,7 @@
 #define _SAFE_CALL_H_
 
 #include <sys/socket.h>
+#include <semaphore.h>
 #include <sys/types.h>
 #include <sys/stat.h>
 #include <unistd.h>
@@ -64,7 +65,7 @@ char * xreadlink_malloc_n(const char *filename, ssize_t *n);
 int xutime(const char *filename, const struct utimbuf *times);
 int xchdir(const char *path);
 int xsocket(int domain, int type, int proto);
-int xsend(int sockfd, const void *buf, size_t len, int flags);
+ssize_t xsend(int sockfd, const void *buf, size_t len, int flags);
 int xaccept(int sockfd, struct sockaddr *addr, socklen_t *addrlen);
 ssize_t xrecv(int sockfd, void *buf, size_t len, int flags);
 int xbind(int sockfd, const struct sockaddr *addr, socklen_t addrlen);
